Move stuff to CGAL/boost/parameter
